requirements.txt
# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Geospatial processing
geopandas>=0.13.0
shapely>=2.0.0
pyproj>=3.4.0

# For Snowflake
snowflake-connector-python[pandas]>=3.0.0

# Environment management
python-dotenv>=1.0.0

# Azure Blob Storage (for GigaSpatial)
azure-storage-blob>=12.0.0

# GigaSpatial platform integration
giga-spatial>=0.7.0
# pip install giga-spatial
psutil>=5.9.0